% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_regression_scatterplot.R
\name{plot_regression_scatterplot}
\alias{plot_regression_scatterplot}
\title{Create scatter plot graph}
\usage{
plot_regression_scatterplot(
  df,
  x_var,
  y_var,
  color_var = NULL,
  line_se = TRUE,
  line_color = "black",
  x_labs = "X Axis Label",
  y_labs = "Y Axis Label",
  title_labs = "Plot Title"
)
}
\arguments{
\item{df}{A tidy dataframe}

\item{x_var}{Unquoted column name to plot on the x-axis}

\item{y_var}{Unquoted column name to plot on the y-axis}

\item{color_var}{\link{Optional} Column name to map a color scheme to data points,}

\item{line_se}{\link{Optional} Display standard error of regression line, default value is True}

\item{line_color}{\link{Optional} Color of regression line, default value is red}

\item{x_labs}{\link{Optional} X-axis label}

\item{y_labs}{\link{Optional} Y-axis label}

\item{title_labs}{\link{Optional} Plot title}

\item{alpha}{\link{Optional} Float from range 0,1 to control regression line transparency, default value is 0.5}
}
\value{
A labelled scatterplot with a simple linear regression line, using a linear smoothing method with
}
\description{
Creates a scatterplot with a regression line, taking in two numeric variables for x-axis and y-axis,
with optional variables for formatting,
To be used with docopt
}
\examples{

plot_regression_scatterplot <- function(df =  data.frame(x=c(10, 20, 30, 40),
                                                         y=c(-10, -50, -100, -150),
                                                         z=c("class1", "class2" "class3", "class4"))
                                               x_var = "x",
                                               y_var = "y",
                                               color_var = "z",
                                               line_se = TRUE,
                                               line_color = "blue",
                                               x_labs = "Label for X-Axis",
                                               y_labs = "Label for Y-Axis",
                                               title_labs = "Plot Title")



}
