% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{Read the input file and split dataset into training and testing sets}
\usage{
split_data(file_path, prop = 0.8, seed = NULL)
}
\arguments{
\item{file_path}{A string path to the input CSV file}

\item{prop}{Proportion of data for training (default = 0.8)}

\item{seed}{Optional integer for reproducibility}
}
\value{
A list with two items.
The first item (named train) contains the training dataset.
The second item (named test) contains the testing dataset.
}
\description{
Read the input file and split dataset into training and testing sets
}
